issue #69 working STOs..
